pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gremlin-qa"
version = "0.1.0"
description = "Exploratory QA agent - surfaces risk scenarios using curated patterns + LLM reasoning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Abhi" }
]
keywords = ["qa", "testing", "cli", "llm", "exploratory-testing", "risk-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
gremlin = "gremlin.cli:app"

[project.urls]
Homepage = "https://github.com/abhi10/gremlin"
Documentation = "https://github.com/abhi10/gremlin#readme"
Repository = "https://github.com/abhi10/gremlin"
Issues = "https://github.com/abhi10/gremlin/issues"

[tool.hatch.build.targets.wheel]
packages = ["gremlin"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
